+++ src/graia/ariadne/console/__init__.py
@@ -3,6 +3,7 @@
 注意, 本实现并不 robust, 但是可以使用
 """
 
+import asyncio
 import contextlib
 import importlib.metadata
 import sys
@@ -10,6 +11,7 @@ from asyncio.events import AbstractEventLoop
 from asyncio.tasks import Task
 from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 
+from creart import it
 from loguru import logger
 from prompt_toolkit.formatted_text import AnyFormattedText
 from prompt_toolkit.patch_stdout import StdoutProxy
@@ -173,7 +175,7 @@ class Console:
                 if interface.annotation is Broadcast:
                     return self.console.broadcast
                 if interface.annotation is AbstractEventLoop:
-                    return self.console.broadcast.loop
+                    return asyncio.get_running_loop()
 
         while self.running:
             try:
@@ -214,7 +216,7 @@ class Console:
                     logger.remove(0)
                 self.handler_id = logger.add(StdoutProxy(raw=True))  # type: ignore
 
-            self.task = self.broadcast.loop.create_task(self.loop())
+            self.task = it(AbstractEventLoop).create_task(self.loop())
 
     def stop(self):
         """提示 Console 停止, 非异步, 幂等"""